stylized PDF download
